+++ mep/books/queryset.py
@@ -10,6 +10,7 @@ class WorkSolrQuerySet(AliasedSolrQuerySet):
 
     #: map readable field names to actual solr fields
     field_aliases = {
+        'pk': 'pk_i',
         'title': 'title_t',
         'sort_title': 'sort_title_isort',
         'authors': 'authors_t',
@@ -22,4 +23,13 @@ class WorkSolrQuerySet(AliasedSolrQuerySet):
         'account_start': 'account_start_i',
         'account_end': 'account_end_i',
         'is_uncertain': 'is_uncertain_b',
+        'event_count': 'event_count_i',
+        'event_years': 'event_years_is',
     }
+
+    # edismax alias for searching on admin work pseudo-field
+    admin_work_qf = '{!qf=$admin_work_qf pf=$admin_work_pf v=$work_query}'
+
+    def search_admin_work(self, search_term):
+        return self.search(self.admin_work_qf) \
+            .raw_query_parameters(work_query=search_term)